requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
selenium==4.15.2
webdriver-manager==4.0.1
gunicorn==21.2.0
python-jose==3.3.0
cryptography==41.0.7
pandas==2.1.4
openpyxl==3.1.2 